+++ lib/Screens/LocalMusic/local_song.dart
@@ -1,23 +1,30 @@
 import 'dart:io';
 import 'package:blackhole/Services/player_service.dart';
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 import 'package:on_audio_query/on_audio_query.dart';
 
-class SongsController extends GetxController {
-  final OnAudioQuery audioQuery = OnAudioQuery();
-  final hasPermission = false.obs;
+class SongsWidget extends StatefulWidget {
+  const SongsWidget({Key? key}) : super(key: key);
 
   @override
-  void onInit() {
-    super.onInit();
+  _SongsWidgetState createState() => _SongsWidgetState();
+}
+
+class _SongsWidgetState extends State<SongsWidget> {
+  final OnAudioQuery _audioQuery = OnAudioQuery();
+  bool _hasPermission = false;
+
+  @override
+  void initState() {
+    super.initState();
     LogConfig logConfig = LogConfig(logType: LogType.DEBUG);
-    audioQuery.setLogConfig(logConfig);
+    _audioQuery.setLogConfig(logConfig);
     checkAndRequestPermissions();
   }
 
-  Future<void> checkAndRequestPermissions({bool retry = false}) async {
-    hasPermission.value = await audioQuery.checkAndRequest(retryRequest: retry);
+  checkAndRequestPermissions({bool retry = false}) async {
+    _hasPermission = await _audioQuery.checkAndRequest(retryRequest: retry);
+    if (_hasPermission) setState(() {});
   }
 
   Future<void> deleteSong(String filePath) async {
@@ -25,86 +32,82 @@ class SongsController extends GetxController {
       final file = File(filePath);
       if (await file.exists()) {
         await file.delete();
-        hasPermission.value = hasPermission.value; // Trigger rebuild
+        setState(() {});
+      } else {
+        print('File not found');
       }
     } catch (e) {
-      // Handle error
+      print('Error deleting file: $e');
     }
   }
-}
-
-class SongsWidget extends StatelessWidget {
-  const SongsWidget({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final controller = Get.put(SongsController());
-
     return Scaffold(
       appBar: AppBar(
         title: const Text('MUSIC'),
         elevation: 2,
       ),
       body: Center(
-        child: Obx(
-          () => !controller.hasPermission.value
-              ? noAccessToLibraryWidget(controller)
-              : FutureBuilder<List<SongModel>>(
-                  future: controller.audioQuery.querySongs(
-                    sortType: SongSortType.TITLE,
-                    orderType: OrderType.ASC_OR_SMALLER,
-                    uriType: UriType.EXTERNAL,
-                    ignoreCase: true,
-                  ),
-                  builder: (context, item) {
-                    if (item.hasError) {
-                      return Text(item.error.toString());
-                    }
-                    if (item.data == null) {
-                      return const CircularProgressIndicator();
-                    }
-                    if (item.data!.isEmpty) return const Text('Nothing found!');
-                    return item.data!.isNotEmpty
-                        ? ListView.builder(
-                            itemCount: item.data!.length,
-                            itemBuilder: (context, index) {
-                              final song = item.data![index];
-                              return ListTile(
-                                title: Text(
-                                  song.title!,
-                                  maxLines: 1,
-                                  overflow: TextOverflow.ellipsis,
-                                ),
-                                subtitle: Text(song.artist ?? 'No Artist'),
-                                leading: QueryArtworkWidget(
-                                  controller: controller.audioQuery,
-                                  id: song.id,
-                                  type: ArtworkType.AUDIO,
-                                  nullArtworkWidget: CircleAvatar(
-                                    child: Icon(Icons.person),
-                                  ),
-                                ),
-                                onTap: () {
-                                  PlayerInvoke.init(
-                                    songsList: item.data!,
-                                    index: index,
-                                    isOffline: true,
-                                    fromDownloads: false,
-                                    recommend: false,
-                                  );
-                                },
-                              );
-                            },
-                          )
-                        : const Center(child: Text('No songs found'));
-                  },
+        child: !_hasPermission
+            ? noAccessToLibraryWidget()
+            : FutureBuilder<List<SongModel>>(
+                future: _audioQuery.querySongs(
+                  sortType: SongSortType.TITLE,
+                  orderType: OrderType.ASC_OR_SMALLER,
+                  uriType: UriType.EXTERNAL,
+                  ignoreCase: true,
                 ),
-        ),
+                builder: (context, item) {
+                  if (item.hasError) {
+                    return Text(item.error.toString());
+                  }
+                  if (item.data == null) {
+                    return const CircularProgressIndicator();
+                  }
+                  if (item.data!.isEmpty) return const Text('Nothing found!');
+                  return item.data!.isNotEmpty
+                      ? ListView.builder(
+                          itemCount: item.data!.length,
+                          itemBuilder: (context, index) {
+                            final song = item.data![index];
+                            print(
+                                'Song title: ${song.title}, Artist: ${song.artist}');
+                            return ListTile(
+                              title: Text(
+                                song.title!,
+                                maxLines: 1,
+                                overflow: TextOverflow.ellipsis,
+                              ),
+                              subtitle: Text(song.artist ?? 'No Artist'),
+                              leading: QueryArtworkWidget(
+                                controller: _audioQuery,
+                                id: song.id,
+                                type: ArtworkType.AUDIO,
+                                nullArtworkWidget: CircleAvatar(
+                                  child: Icon(Icons.person),
+                                ),
+                              ),
+                              onTap: () {
+                                PlayerInvoke.init(
+                                  songsList: item.data!,
+                                  index: index,
+                                  isOffline: true,
+                                  fromDownloads: false,
+                                  recommend: false,
+                                );
+                              },
+                            );
+                          },
+                        )
+                      : const Center(child: Text('No songs found'));
+                },
+              ),
       ),
     );
   }
 
-  Widget noAccessToLibraryWidget(SongsController controller) {
+  Widget noAccessToLibraryWidget() {
     return Container(
       decoration: BoxDecoration(
         borderRadius: BorderRadius.circular(10),
@@ -117,7 +120,7 @@ class SongsWidget extends StatelessWidget {
           const Text("Application doesn't have access to the library"),
           const SizedBox(height: 10),
           ElevatedButton(
-            onPressed: () => controller.checkAndRequestPermissions(retry: true),
+            onPressed: () => checkAndRequestPermissions(retry: true),
             child: const Text('Allow'),
           ),
         ],
